requirements.txt
cryptography
playwright
pyautogui
python-dotenv
pydantic
PyYAML
requests
colorama
pytest
